requirements.txt
# For running the web services
fastapi
uvicorn
httpx
python-multipart

# For connecting to PostgreSQL
sqlalchemy
psycopg2-binary

# For Embedder Service
torch
open_clip_torch
pymilvus
Pillow

